#16 #12 #17 Walidation password ,submit donation of gifts

+++ src/main/java/pl/coderslab/charity/service/UserService.java
@@ -1,7 +1,10 @@
 package pl.coderslab.charity.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import pl.coderslab.charity.config.UserInSessionService;
 import pl.coderslab.charity.entity.Donation;
 import pl.coderslab.charity.entity.authentication.Role;
 import pl.coderslab.charity.entity.authentication.User;
@@ -14,11 +17,52 @@ import java.util.Optional;
 public class UserService {
 
     @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
     UserRepository userRepository;
 
+    @Autowired
+    UserInSessionService userInSessionService;
+
+    @Secured("ROLE_ADMIN")
+    public User block(User user) {
+
+        user.setRetypePassword(user.getPassword());
+        if (user.getEnabled() == 1) {
+            user.setEnabled(0);
+        } else {
+            user.setEnabled(1);
+        }
+        return userRepository.save(user);
+
+    }
+
+    public User editPassword(User user){
+        User userFromDatabase = userRepository.findById(user.getHiddenId());
+        String encodedPassword = passwordEncoder.encode(user.getPassword());
+        userFromDatabase.setPassword(encodedPassword);
+        return   userRepository.save(userFromDatabase);
+
+    }
+
+
+    public User editUserInfo(User user) {
+        User userFromDatabase = userRepository.findById(user.getHiddenId());
+        userFromDatabase.setFirstName(user.getFirstName());
+        userFromDatabase.setSecondName(user.getSecondName());
+        userFromDatabase.setEmail(user.getEmail());
+        return userRepository.save(userFromDatabase);
+    }
+
     public User save(User user) {
+        String encodedPassword = passwordEncoder.encode(user.getPassword());
+        user.setPassword(encodedPassword);
+        user.setRetypePassword(encodedPassword);
+        user.setEnabled(1);
         return userRepository.save(user);
     }
+
 
     public User findByEmail(String email) {
         return userRepository.findFirstByEmail(email);
@@ -32,21 +76,26 @@ public class UserService {
         return userRepository.findAll();
     }
 
+    @Secured("ROLE_ADMIN")
     public void delete(User user) {
-        userRepository.delete(user);
+        if (!user.getEmail().equals(userInSessionService.getUserFromSessionByLogin())) {
+            userRepository.delete(user);
+        }
+
+
     }
 
     public User findById(int id) {
-       return userRepository.findById(id);
+        return userRepository.findById(id);
     }
 
-    public List<User> findUsersByRolesContains(Role role){
-       return userRepository.findUsersByRolesContains(role);
-    }
-    public List<User> findUsersByRolesNotContains(Role role){
-       return userRepository.findUsersByRolesNotContains(role);
+    public List<User> findUsersByRolesContains(Role role) {
+        return userRepository.findUsersByRolesContains(role);
     }
 
+    public List<User> findUsersByRolesNotContains(Role role) {
+        return userRepository.findUsersByRolesNotContains(role);
+    }
 
 
 }
